+++ core/src/de/homelab/madgaksha/lotsofbs/util/interpolator/LogDoubleInterpolator.java
@@ -37,8 +37,8 @@ public class LogDoubleInterpolator extends AInterpolator<Double> {
 	@Override
 	protected void doSetup(Object options) {
 		ILogExpInterpolatorOptions o = (ILogExpInterpolatorOptions) options;
-		double a = (double) start;
-		double b = (double) end;
+		double a = start;
+		double b = end;
 		double x1 = o.getX1();
 		double x2 = o.getX2();
 		k1 = (b - a) / (Math.log(x2) - Math.log(x1));
